requirements.txt
Django
django-annoying==0.8.1
django-compressor==1.5
django-cors-headers==1.1.0
bpython==0.14.1
Unipath==1.1

djangorestframework==3.6.3
django-filter==1.0.4
tweepy==3.5.0
